requirements.txt
kafka-python>=2.0.1
Flask>=1.1.1